backend-flask/requirements.txt
pip==23.0.1

flask
flask-cors
python-dotenv

# honeycomb.io
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests

# CloudWatch Logs
# watchtower
